requirements.txt
torch
numpy
tensorflow
pytest
matplotlib
hydra-core
